+++ application/modules/ms/controllers/provinsi.php
@@ -27,15 +27,30 @@ class provinsi extends CI_Controller {
 	}
 	
 	public function index(){
-		/**
-		 * FIXME
-		 * this only temporary set list object data
-		 * please switch to ajax request to datatable() method if this page has rendered
-		 */
+		$page   = (!$this->input->get('page')) ? 1 : $this->input->get('page');
 
-		$this->twiggy->set('list', ProvinsiEntity::all());
-
+		$this->twiggy->set('this_page', $page);
 		$this->twiggy->template('master/provinsi/index')->display();
+	}
+
+	public function list_data()
+	{
+		// get pagable data
+		$page   = (!$this->input->get('page')) ? 1 : $this->input->get('page');
+		$limit  = 3;
+		$offset = (($page-1)*$limit);
+		$search = "";
+
+		$this->model = new ProvinsiEntity();
+		$list = $this->model->get_provinsi($offset, $limit, $search, null, null);
+		$total = $this->model->get_provinsi_count($search);
+
+		$this->twiggy->set('list', $list->result());
+		$this->twiggy->set('total', $total);
+		$this->twiggy->set('totalPage', ceil($total/$limit));
+		$this->twiggy->set('size', $list->num_rows());
+		$this->twiggy->set('page', $page);
+		$this->twiggy->template('master/provinsi/list')->display();
 	}
 
 	public function form($id=null){
@@ -48,13 +63,16 @@ class provinsi extends CI_Controller {
 	}
 
 	public function delete($id){
-		if ($id == null) {
+		if($this->input->is_ajax_request())
+		{
+			if ($id == null) {
+				redirect($this->direct, 'location', 303);
+			}
+
+			ProvinsiEntity::delete($id);
+
 			redirect($this->direct, 'location', 303);
 		}
-
-		ProvinsiEntity::delete($id);
-
-		redirect($this->direct, 'location', 303);
 	}
 
 	public function submit(){
